move jest matchers and their tests to /individual folder
