+++ server/test/test_utils/matchers/serviceError/individual/toBeSE_NOT_ARRAY.js
@@ -1,7 +1,7 @@
-import { SEReason } from "../../../../router/api/v2/routeBuilder/core/service/dataExtractors/error/SEReason";
-import { ServiceError } from "../../../../router/api/v2/routeBuilder/core/service/dataExtractors/error/ServiceError";
-import MatcherReturner from "../../jest_util/MatcherReturner";
-import { isErrorWithReason } from "./isErrorWithReason";
+import { SEReason } from "../../../../../router/api/v2/routeBuilder/core/service/dataExtractors/error/SEReason";
+import { ServiceError } from "../../../../../router/api/v2/routeBuilder/core/service/dataExtractors/error/ServiceError";
+import MatcherReturner from "../../../jest_util/MatcherReturner";
+import { isErrorWithReason } from "../isErrorWithReason";
 
 /**
  * Jest matcher checks whenever provided param is ServiceError with reason NOT_ARRAY
