+++ server/test/test_utils/test/matchers/serviceError/individual/toBeSE_MISCONFIGURED.test.js
@@ -1,7 +1,7 @@
-import { SEReason } from "../../../../../router/api/v2/routeBuilder/core/service/dataExtractors/error/SEReason";
-import { ServiceError } from "../../../../../router/api/v2/routeBuilder/core/service/dataExtractors/error/ServiceError";
-import passJestThis from "../../../jest_util/passJestThisObject";
-import { toBeSE_MISCONFIGURED } from "../../../matchers/serviceError/toBeSE_MISCONFIGURED";
+import { SEReason } from "../../../../../../router/api/v2/routeBuilder/core/service/dataExtractors/error/SEReason";
+import { ServiceError } from "../../../../../../router/api/v2/routeBuilder/core/service/dataExtractors/error/ServiceError";
+import passJestThis from "../../../../jest_util/passJestThisObject";
+import { toBeSE_MISCONFIGURED } from "../../../../matchers/serviceError/individual/toBeSE_MISCONFIGURED";
 
 describe('toBeSE_MISCONFIGURED() test suite', () => {  
     it('Should return object with pass equal to true if the object is ServiceError MISCONFIGURED', () => {
