+++ server/test/test_utils/test/matchers/serviceError/individual/toBeSE_NOT_ALLOWED.test.js
@@ -1,7 +1,7 @@
-import { SEReason } from "../../../../../router/api/v2/routeBuilder/core/service/dataExtractors/error/SEReason";
-import { ServiceError } from "../../../../../router/api/v2/routeBuilder/core/service/dataExtractors/error/ServiceError";
-import passJestThis from "../../../jest_util/passJestThisObject";
-import { toBeSE_NOT_ALLOWED } from "../../../matchers/serviceError/toBeSE_NOT_ALLOWED";
+import { SEReason } from "../../../../../../router/api/v2/routeBuilder/core/service/dataExtractors/error/SEReason";
+import { ServiceError } from "../../../../../../router/api/v2/routeBuilder/core/service/dataExtractors/error/ServiceError";
+import passJestThis from "../../../../jest_util/passJestThisObject";
+import { toBeSE_NOT_ALLOWED } from "../../../../matchers/serviceError/individual/toBeSE_NOT_ALLOWED";
 
 describe('toBeSE_NOT_ALLOWED() test suite', () => {
     it('Should return object with pass equal to true if the object is ServiceError NOT_ALLOWED', () => {
